everything works

+++ gui.py
@@ -7,14 +7,7 @@ from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg, NavigationToolb
 from pandas import DataFrame
 from datetime import datetime
 from FunctionBuilder import FunctionBuilder
-
-#TODO
-#skompilowac do exe
-#obsluga klawiatury na dolne wiersze
-#wyjątki
-#usuwanie wykresu dla nowych danych
-#opakowac w funkcje
-#naprawić epsilony
+from tkinter import messagebox
 
 FONT_SIZE = 12
 FONT_STYLE = "Default"
@@ -36,6 +29,9 @@ class Gui:
         self.label_n = None
         self.file_to_save = None
         self.is_custom = False
+        self.exist_chart = False
+        self.toolbar_exist = False
+        self.picked_checkbox = False
 
         self.n = 2
 
@@ -45,9 +41,9 @@ class Gui:
         label_t0 = Label(self.input_frame,text="t\u2080= ",font=(FONT_STYLE, FONT_SIZE))
         label_tn = Label(self.input_frame,text="t\u2099= ",font=(FONT_STYLE, FONT_SIZE))
 
-        label_y.grid(column=1, row=0, pady=5, padx=10,sticky=E)
-        label_t0.grid(column=1, row=1, pady=5, padx=10,sticky=E)
-        label_x0.grid(column=1, row=2, pady=5, padx=10,sticky=W)
+        label_y.grid(column=1, row=0, pady=5,sticky=E)
+        label_t0.grid(column=1, row=1, pady=5,sticky=E)
+        label_x0.grid(column=1, row=2, pady=5,sticky=W)
 
 
         self.input_equation_entry.grid(column=2, row=0, pady=5)
@@ -66,9 +62,14 @@ class Gui:
 
 
     def _button_click(self, char):
-        current = self.input_equation_entry.get()
-        self.input_equation_entry.delete(0, END)
-        self.input_equation_entry.insert(0, str(current) + str(char))
+        inputs = [self.input_equation_entry,self.input_tn,self.input_x0,self.input_t0]
+
+
+        for entry in inputs:
+            if entry.focus():
+                current = entry.get()
+                entry.delete(0, END)
+                entry.insert(0, str(current) + str(char))
 
     def _backspace_click(self):
         current = self.input_equation_entry.get()
@@ -80,60 +81,16 @@ class Gui:
     def _create_keyboard(self):
         keyboard = Frame(self.gui)
 
-        button_0 = Button(keyboard, text="0", padx=20, pady=20, command=lambda: self._button_click(0))
-        button_1 = Button(keyboard, text="1", padx=20, pady=20, command=lambda: self._button_click(1))
-        button_2 = Button(keyboard, text="2", padx=20, pady=20, command=lambda: self._button_click(2))
-        button_3 = Button(keyboard, text="3", padx=20, pady=20, command=lambda: self._button_click(3))
-        button_4 = Button(keyboard, text="4", padx=20, pady=20, command=lambda: self._button_click(4))
-        button_5 = Button(keyboard, text="5", padx=20, pady=20, command=lambda: self._button_click(5))
-        button_6 = Button(keyboard, text="6", padx=20, pady=20, command=lambda: self._button_click(6))
-        button_7 = Button(keyboard, text="7", padx=20, pady=20, command=lambda: self._button_click(7))
-        button_8 = Button(keyboard, text="8", padx=20, pady=20, command=lambda: self._button_click(8))
-        button_9 = Button(keyboard, text="9", padx=20, pady=20, command=lambda: self._button_click(9))
-
-        button_t = Button(keyboard, text="t", padx=20, pady=20, command=lambda: self._button_click("t"))
-        button_xt = Button(keyboard, text="x(t)", padx=20, pady=20, command=lambda: self._button_click("x(t)"))
-
-        button_plus = Button(keyboard, text="+", padx=20, pady=20, command=lambda: self._button_click("+"))
-        button_minus = Button(keyboard, text="-", padx=20, pady=20, command=lambda: self._button_click("-"))
-        button_division = Button(keyboard, text="/", padx=20, pady=20, command=lambda: self._button_click("/"))
-        button_multiplication = Button(keyboard, text="*", padx=20, pady=20, command=lambda: self._button_click("*"))
-        button_sinus = Button(keyboard, text="sin", padx=20, pady=20,
-                              command=lambda: self._button_click("sin"))  # tutaj mozna dodac otwieranie nawiasu
-        button_cosinus = Button(keyboard, text="cos", padx=20, pady=20, command=lambda: self._button_click("cos"))
-        button_open_bracket = Button(keyboard, text="(", padx=20, pady=20, command=lambda: self._button_click("("))
-        button_close_bracket = Button(keyboard, text=")", padx=20, pady=20, command=lambda: self._button_click(")"))
-        button_backspace = Button(keyboard, text="backspace", padx=20, pady=20, command=self._backspace_click)
         button_more_steps = Button(keyboard, text="next step",padx=20, pady=20, command=self._update_chart)
-
-        button_0.grid(column=0, row=0)
-        button_1.grid(column=1, row=0)
-        button_2.grid(column=2, row=0)
-        button_3.grid(column=3, row=0)
-        button_4.grid(column=4, row=0)
-        button_5.grid(column=5, row=0)
-        button_6.grid(column=6, row=0)
-        button_7.grid(column=7, row=0)
-        button_8.grid(column=8, row=0)
-        button_9.grid(column=9, row=0)
-        button_backspace.grid(column=10, row=0)
-
-        button_t.grid(column=0, row=1)
-        button_xt.grid(column=1, row=1)
-        button_plus.grid(column=2, row=1)
-        button_minus.grid(column=3, row=1)
-        button_division.grid(column=4, row=1)
-        button_multiplication.grid(column=5, row=1)
-        button_sinus.grid(column=6, row=1)
-        button_cosinus.grid(column=7, row=1)
-        button_open_bracket.grid(column=8, row=1)
-        button_close_bracket.grid(column=9, row=1)
-        button_more_steps.grid(column=10, row=1)
-
+        button_more_steps.grid(column=0, row=0)
 
         keyboard.pack(side=BOTTOM)
 
     def _update_chart(self):
+        if not self.exist_chart:
+            self._show_error_message_no_calculate()
+            return
+
         self.n *= 2
         self.figure.clear()
         self.label_n.destroy()
@@ -143,8 +100,7 @@ class Gui:
             data = RungeKutta.runge_kutta_custom(float(self.input_t0.get()), float(self.input_x0.get()),
                                                  float(self.input_tn.get()), self.n,
                                                  FunctionBuilder(self.input_equation_entry.get()).function)
-            epsilon = 0
-
+            self._show_statistics()
         else:
             self.label_epsilon.destroy()
             data, solution, epsilon = RungeKutta.runge_kutta(float(self.input_t0.get()), float(self.input_x0.get()),
@@ -154,44 +110,19 @@ class Gui:
             df_solution = DataFrame(solution, columns=['exact_t', 'exact_x'])
             df_solution = df_solution[['exact_t', 'exact_x']].groupby('exact_t').sum()
             df_solution.plot(kind='line', legend=True, ax=ax, color='r', marker=',', fontsize=10)
-
-
-        df = DataFrame(data,columns=['t','x(t)'])
-        df = df[['t','x(t)']].groupby('t').sum()
-        df.plot(kind='line', legend=True, ax=ax, color='b',marker='o', fontsize=10)
-
-
-
-        ax.set_title('Solution plot')
-        self._show_statistics(epsilon)
-        self.chart_type.draw_idle()
-
-    def _create_chart(self, data, solution):
-        self.figure = plt.Figure(figsize=(7, 4), dpi=80)
-        ax = self.figure.add_subplot(111)
-        self.chart_type = FigureCanvasTkAgg(self.figure, self.gui)
-        self.chart_type.get_tk_widget().pack(side=BOTTOM)
-
-        toolbar = NavigationToolbar2Tk(self.chart_type,self.gui)
-        toolbar.update()
-        toolbar.pack(side=BOTTOM, fill=Y)
+            self._show_statistics(epsilon)
 
         df = DataFrame(data,columns=['t','x(t)'])
         self.file_to_save = df
         df = df[['t','x(t)']].groupby('t').sum()
         df.plot(kind='line', legend=True, ax=ax, color='b',marker='o', fontsize=10)
-
-        df_solution = DataFrame(solution,columns=['exact_t','exact_x'])
-        df_solution = df_solution[['exact_t','exact_x']].groupby('exact_t').sum()
-        df_solution.plot(kind='line', legend=True, ax=ax, color='r',marker=',', fontsize=10)
-
-
-
         ax.set_title('Solution plot')
+
+        self.chart_type.draw_idle()
 
     def _create_checkboxes(self):
         var = StringVar(self.input_frame, "0")
-        label_functions = Label(self.input_frame,text="Functions: ",pady=5,font=(FONT_STYLE, FONT_SIZE))
+        label_functions = Label(self.input_frame,text="Functions: ",pady=5,padx=10,font=(FONT_STYLE, FONT_SIZE))
         label_functions.grid(column=0,row=0,sticky=W)
 
         checkboxes = {"x' = -2 * x * e\u1D57" : "1","x' = (t - 3/t + 2) * (1/2)" : "2","x' = (x\u00B2 + t\u00B2)/(t*x)" : "3"}
@@ -199,15 +130,17 @@ class Gui:
         cnt = 1
         for (text, value) in checkboxes.items():
             Radiobutton(self.input_frame, text=text, variable=var,
-                        value=value,font=(FONT_STYLE, FONT_SIZE),command=lambda:self._handle_checkboxes(var)).grid(column=0,row=cnt, ipady=5, sticky=W)
+                        value=value,font=(FONT_STYLE, FONT_SIZE),command=lambda:self._handle_checkboxes(var),
+                        padx=10).grid(column=0,row=cnt, pady=5, sticky=W)
             cnt += 1
 
         Radiobutton(self.input_frame, text="custom", variable=var,
-                    value="4", font=(FONT_STYLE, FONT_SIZE), command=lambda: self._handle_checkboxes(var)).grid(
-            column=0, row=cnt, ipady=5, sticky=W)
+                    value="4", font=(FONT_STYLE, FONT_SIZE), command=lambda: self._handle_checkboxes(var),padx=10).grid(
+            column=0, row=cnt, pady=5, sticky=W)
 
 
     def _handle_checkboxes(self,var):
+        self.picked_checkbox = True
         options = [["-2 * x * e**t",0,1,1],["(t - 3/t + 2) * (1/2)",1,2,0],["(x**2 + t**2)/(t*x)",1,2,-1]] #f,t0,tn,x0
         self.input_equation_entry.delete(0,END)
         self.input_x0.delete(0,END)
@@ -219,6 +152,8 @@ class Gui:
         if ind == 3:
             self.is_custom = True
             return
+        else:
+            self.is_custom = False
 
         self.input_equation_entry.insert(0,options[ind][0])
         self.input_t0.insert(0,options[ind][1])
@@ -230,23 +165,81 @@ class Gui:
         self.function = functions[ind]
 
 
+
     def return_parameters(self):
         return self.input_equation_entry.get(),self.input_t0.get(),self.input_tn.get(),self.input_x0.get()
 
 
-    def _show_statistics(self,epsilon):
+    def _show_statistics(self,epsilon=None):
         self.label_n = Label(self.gui,text=f"steps: {str(self.n)}",font=(FONT_STYLE, FONT_SIZE))
-        self.label_epsilon = Label(self.gui,text=f"epsilon: {str(epsilon)}",font=(FONT_STYLE, FONT_SIZE))
-
         self.label_n.pack()
-        self.label_epsilon.pack()
+
+
+        if not self.is_custom:
+            self.label_epsilon = Label(self.gui, text=f"epsilon: {str(epsilon)}", font=(FONT_STYLE, FONT_SIZE))
+            self.label_epsilon.pack()
+
+    def _create_chart(self, data, solution=None):
+        self.figure = plt.Figure(figsize=(7, 4.8), dpi=80)
+        ax = self.figure.add_subplot(111)
+        self.chart_type = FigureCanvasTkAgg(self.figure, self.gui)
+
+        if not self.toolbar_exist:
+            toolbar = NavigationToolbar2Tk(self.chart_type,self.gui)
+            toolbar.update()
+            toolbar.pack(side=BOTTOM, fill=Y)
+            self.toolbar_exist = True
+
+        self.chart_type.get_tk_widget().pack(side=BOTTOM)
+
+        df = DataFrame(data,columns=['t','x(t)'])
+        self.file_to_save = df
+        df = df[['t','x(t)']].groupby('t').sum()
+        df.plot(kind='line', legend=True, ax=ax, color='b',marker='o', fontsize=10)
+
+        if not self.is_custom:
+            df_solution = DataFrame(solution,columns=['exact_t','exact_x'])
+            df_solution = df_solution[['exact_t','exact_x']].groupby('exact_t').sum()
+            df_solution.plot(kind='line', legend=True, ax=ax, color='r',marker=',', fontsize=10)
+
+
+        ax.set_title('Solution plot')
+        self.exist_chart = True
+
+    def _reset(self):
+        self.n = 2
+        self.figure.clear()
+        self.chart_type.get_tk_widget().destroy()
+        self.label_n.destroy()
+        self.label_epsilon.destroy()
+
 
     def _calculate(self):
+        for entry in self.return_parameters():
+            if len(entry) == 0:
+
+                if self.exist_chart:
+                    self.exist_chart = False
+                    self._reset()
+
+                self._show_error_message_no_data()
+                return
+
+        if self.exist_chart is True:
+            self._reset()
+
+        if not self.picked_checkbox:
+            self._show_error_message_no_checkbox()
+            return
+
+        self.label_n = Label(self.gui, text=f"steps: {str(self.n)}", font=(FONT_STYLE, FONT_SIZE))
+        self.label_n.pack()
+
         if self.is_custom:
             data = RungeKutta.runge_kutta_custom(float(self.input_t0.get()), float(self.input_x0.get()),
                                                 float(self.input_tn.get()), self.n,
                                           FunctionBuilder(self.input_equation_entry.get()).function)
-            self._create_custom_chart(data)
+            self._create_chart(data)
 
         else:
             data, solution,epsilon = RungeKutta.runge_kutta(float(self.input_t0.get()), float(self.input_x0.get()),
@@ -255,33 +248,26 @@ class Gui:
             self.label_epsilon.pack()
             self._create_chart(data, solution)
 
-        self.label_n = Label(self.gui, text=f"steps: {str(self.n)}", font=(FONT_STYLE, FONT_SIZE))
-        self.label_n.pack()
-
-    def _create_custom_chart(self,data):
-        self.figure = plt.Figure(figsize=(7, 4), dpi=80)
-        ax = self.figure.add_subplot(111)
-        self.chart_type = FigureCanvasTkAgg(self.figure, self.gui)
-        self.chart_type.get_tk_widget().pack(side=BOTTOM)
-
-        toolbar = NavigationToolbar2Tk(self.chart_type,self.gui)
-        toolbar.update()
-        toolbar.pack(side=BOTTOM, fill=Y)
-
-        df = DataFrame(data,columns=['t','x(t)'])
-        self.file_to_save = df
-        df = df[['t','x(t)']].groupby('t').sum()
-        df.plot(kind='line', legend=True, ax=ax, color='b',marker='o', fontsize=10)
-
-        ax.set_title('Solution plot')
-
     def _create_save_button(self):
         save_button = Button(self.input_frame, text="save", command=self._save, width=17,
                               font=(FONT_STYLE, FONT_SIZE))
         save_button.grid(column=2,row=3,pady=5,sticky=E)
 
     def _save(self):
+        if not self.exist_chart:
+            self._show_error_message_no_calculate()
+            return
+
         self.file_to_save.to_csv("data " + str(datetime.now().strftime("%H_%M_%S") + ".txt"), encoding='utf-8', index=False, sep=';')
+
+    def _show_error_message_no_data(self):
+        messagebox.showerror('Error', 'There is no input!')
+
+    def _show_error_message_no_calculate(self):
+        messagebox.showerror('Error', 'No result!')
+
+    def _show_error_message_no_checkbox(self):
+        messagebox.showerror('Error', 'Select functions!')
 
     def run(self):
         self._create_input_frame()
